Changed id type of data representing structs to `u64` `Unsigned<BigInt>` is the default id type `diesel migration run` generates for `schema.rs` (See: http://localhost:8000/users/test). Also fixed postman script (missing requests due export).

+++ src/auth/auth.rs
@@ -6,7 +6,7 @@ use rocket::request::{FromRequest, Outcome};
 use rocket::Request;
 
 pub struct AuthenticatedUser {
-    user_id: i32,
+    user_id: u64,
 }
 
 #[derive(Debug)]
